requirements36_v2.txt
adabound>=0.0.5
alabaster==0.7.12
albumentations==1.0.2
antlr4-python3-runtime==4.8
astunparse==1.6.3
babel==2.8.0
beautifulsoup4==4.10.0
chumpy==0.70
compress-pickle==1.2.0
configparser==5.0.1
deepspeech-gpu==0.9.2
docker-pycreds==0.4.0
docutils==0.16
dominate==2.6.0
easydict==1.9
face-alignment>=1.3.3
facenet-pytorch==2.5.1
filelock==3.0.12
flatbuffers==1.12
freetype-py==2.2.0
gdown==4.2.1
gitdb==4.0.5
gitpython==3.1.11
google-pasta==0.2.0
hydra-core==1.0.5
imagesize==1.2.0
imgaug==0.4.0
importlib-resources==5.1.0
insightface==0.5
install==1.3.5
itk-core==5.1.2
itk-filtering==5.1.2
itk-meshtopolydata==0.6.2
itk-numerics==5.1.2
itk-registration==5.1.2
itk-segmentation==5.1.2
kaggle==1.5.12
keras-preprocessing==1.1.2
kornia==0.6.1
lightning-bolts>=0.4.0
matplotlib==3.3.4
mediapipe
memory-profiler==0.58.0
mmcv==1.2.2
munch==2.5.0
#- numpy>=1.19.3
# nvdiffrast==0.2.7
omegaconf==2.0.6
onnx==1.10.2
onnxruntime-gpu==1.9.0
#- opencv-contrib-python==4.5.5.62
#- opencv-python==4.5.5.62
#- opencv-python-headless==4.5.5.62
opt-einsum==3.3.0
pathtools==0.1.2
pillow==8.1.2
prettytable==2.4.0
promise==2.3
protobuf==3.13.0
#psbody-mesh==0.4
psutil==5.7.3
pyasn1-modules==0.2.8
pydeprecate==0.3.1
pyglet==1.5.11
PyOpengl==3.1.5
- # pyrender==0.1.39
python-slugify==5.0.2
python-speech-features==0.6
pytorch-lightning==1.4.9
#- # pytorch3d==0.6.0
#- # pytorch3d==0.6.1
pytorch3d -f https://dl.fbaipublicfiles.com/pytorch3d/packaging/wheels/py37_cu111_pyt191/download.html
pyzmq==22.0.3
# scikit-sound==0.2.3
#- scikit-video==1.1.11
scipy==1.5.4
sentry-sdk==0.19.2
setuptools==54.2.0
shapely==1.7.1
shortuuid==1.0.1
smmap==3.0.4
smplx==0.1.26
snowballstemmer==2.0.0
soupsieve==2.3.1
subprocess32==3.5.4
#tb-nightly==2.5.0a20201210
#tensorboard-data-server==0.6.1
#tensorflow-estimator==2.3.0
text-unidecode==1.3
#tf-estimator-nightly==2.5.0.dev2020120901
#tf-nightly==2.5.0.dev20201210
tensorflow==2.5.0
timm>=0.4.9
torchaudio>=0.7.2
torchfile==0.1.0
torchgeometry==0.1.2
torchmetrics==0.5.1
torchvision==0.10.1
trimesh==3.6.18
wandb==0.10.30
watchdog==0.10.3
wrapt==1.12.1
yapf==0.30.0
youtube-dl==2021.6.6